return the response body as a unicode object
Keep the symmetry with set_request_body() which encodes the unicode
object (Robot Framework defaults to unicode for all strings) into an
instance of str by decoding the response into an instance of unicode in
get_response_body().

Failure to do so causes UnicodeDecodeErrors in all keywords that compare
the response body against a given string if any of the two contains
Unicode characters.